+++ OUDS/Core/Themes/Orange/Sources/OrangeTheme.swift
@@ -133,7 +133,7 @@ import OUDSTokensSemantic
 ///
 /// - Core OUDS version: 1.10.0
 /// - Core Orange version: 1.2.0
-/// - Brand Orange version: 2.4.0
+/// - Brand Orange version: 2.5.0
 ///
 /// - Since: 0.8.0
 open class OrangeTheme: OUDSTheme, @unchecked Sendable {