+++ OUDS/Core/Themes/Orange/Tests/Values/ComponentTokens/MockTheme/MockTheme+AllTextAreaComponentTokens.swift
@@ -36,6 +36,8 @@ final class MockThemeTextAreaComponentTokenProvider: OrangeThemeTextAreaComponen
 
     override var spacePaddingBlock: SpaceSemanticToken { Self.mockThemeTextAreaSpace }
     override var spacePaddingBlockTopEmpty: SpaceSemanticToken { Self.mockThemeTextAreaSpace }
+    override var spacePaddingBlockTrailingContainer: SpaceSemanticToken { Self.mockThemeTextAreaSpace }
+    override var spacePaddingBlockEmptyTrailingContainer: SpaceSemanticToken { Self.mockThemeTextAreaSpace }
 
     // MARK: - Sizes