+++ OUDS/Core/Themes/Wireframe/Sources/WireframeTheme.swift
@@ -51,7 +51,7 @@ import SwiftUI
 ///
 /// - Core OUDS version: 1.10.0
 /// - Core Wireframe version: 1.4.0
-/// - Brand Wireframe version: 2.4.0
+/// - Brand Wireframe version: 2.5.0
 ///
 /// - Since: 0.17.0
 public final class WireframeTheme: OUDSTheme, @unchecked Sendable {