+++ OUDS/Core/Tokens/ComponentTokens/Sources/Values/TextAreaComponentTokens.swift
@@ -24,8 +24,12 @@ import OUDSTokensSemantic
 ///
 /// - Since: 0.17.0
 public protocol TextAreaComponentTokens {
+
     var spacePaddingBlock: SpaceSemanticToken { get }
     var spacePaddingBlockTopEmpty: SpaceSemanticToken { get }
+    var spacePaddingBlockTrailingContainer: SpaceSemanticToken { get }
+    var spacePaddingBlockEmptyTrailingContainer: SpaceSemanticToken { get }
+
     var sizeMinHeightInput: SizeSemanticToken { get }
     var sizeMaxHeightInput: SizeSemanticToken { get }
     var sizeMaxHeightAssetsContainer: SizeSemanticToken { get }